backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
websockets==12.0

# AI and ML
requests==2.31.0
transformers==4.37.0

# Browser Automation
playwright==1.41.0
aiofiles==23.2.1

# Vision and OCR
pytesseract==0.3.10
Pillow

# Database
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
